requirements.txt
# packages in environment at C:\Users\agradillas\Miniconda3\envs\convertAria:
#
# Name                    Version                   Build  Channel
bzip2                     1.0.8                he774522_0  
ca-certificates           2022.4.26            haa95532_0    anaconda
certifi                   2022.6.15       py310haa95532_0    anaconda
cffi                      1.15.1          py310hcbf9ad4_0    conda-forge
cryptography              37.0.4          py310ha857299_0    conda-forge
dnspython                 2.2.1              pyhd8ed1ab_0    conda-forge
greenlet                  1.1.1           py310hd77b12b_0    anaconda
idna                      3.4                pyhd8ed1ab_0    conda-forge
libffi                    3.4.2                hd77b12b_4  
libprotobuf               3.20.1               h12be248_4    conda-forge
libzlib                   1.2.12               hcfcfb64_3    conda-forge
mysql-common              8.0.30               hd2cd81a_1    conda-forge
mysql-connector-python    8.0.30          py310h21f64ac_0    conda-forge
mysql-libs                8.0.30               h4d1747d_1    conda-forge
openssl                   1.1.1q               h2bbff1b_0  
pip                       22.1.2          py310haa95532_0  
protobuf                  3.20.1          py310h5588dad_0    conda-forge
pycparser                 2.21               pyhd8ed1ab_0    conda-forge
python                    3.10.4               hbb2ffb3_0  
python-decouple           3.6                pyhd8ed1ab_0    conda-forge
python_abi                3.10                    2_cp310    conda-forge
setuptools                63.4.1          py310haa95532_0  
six                       1.16.0             pyh6c4a22f_0    conda-forge
sqlalchemy                1.4.32          py310h2bbff1b_0    anaconda
sqlite                    3.39.2               h2bbff1b_0  
tk                        8.6.12               h2bbff1b_0  
tzdata                    2022c                h04d1e81_0  
ucrt                      10.0.20348.0         h57928b3_0    conda-forge
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.29.30139          h890b9b1_7    conda-forge
wheel                     0.37.1             pyhd3eb1b0_0  
wincertstore              0.2             py310haa95532_2  
xz                        5.2.5                h8cc25b3_1  
zlib                      1.2.12               h8cc25b3_3  
zstd                      1.5.2                h7755175_4    conda-forge
